requirements.txt
streamlit==1.51.0
pandas==2.2.2
numpy==1.26.4

sqlalchemy==1.4.54
pyodbc==5.1.0

scikit-learn==1.5.2
plotly==5.24.1

openai==1.56.0
python-dotenv==1.0.1
